+++ src/gemseo_umdo/monte_carlo_sampler.py
@@ -16,8 +16,8 @@
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from typing import TYPE_CHECKING
-from typing import Callable
 
 from gemseo.algos.doe.openturns.openturns import OpenTURNS
 from gemseo.typing import RealArray