+++ tests/statistics/conftest.py
@@ -15,11 +15,12 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from typing import Callable
 
 import pytest
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from numpy import ndarray