+++ tests/statistics/mlmc_mlcv/pilots/test_mean.py
@@ -75,5 +75,5 @@ def test_compute_V_l_delta(  # noqa: N802
     delta = pilot._Mean__delta
     assert_almost_equal(V_l, array([nan, expected_V_l, nan]))
     assert len(delta) == 3
-    for x, y in zip(delta, [array([]), array(expected_x), array([])]):
+    for x, y in zip(delta, [array([]), array(expected_x), array([])], strict=False):
         assert_almost_equal(x, y)