+++ tests/visualizations/test_uncertain_coupling_graph.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 import pytest
 from gemseo.algos.design_space import DesignSpace
@@ -22,13 +23,15 @@ from gemseo.disciplines.analytic import AnalyticDiscipline
 from gemseo.post._graph_view import GraphView
 from gemseo.problems.mdo.sobieski.core.problem import SobieskiProblem
 from gemseo.problems.mdo.sobieski.disciplines import SobieskiAerodynamics
-from gemseo.problems.mdo.sobieski.disciplines import SobieskiDiscipline
 from gemseo.problems.mdo.sobieski.disciplines import SobieskiMission
 from gemseo.problems.mdo.sobieski.disciplines import SobieskiPropulsion
 from gemseo.problems.mdo.sobieski.disciplines import SobieskiStructure
 from gemseo.utils.data_conversion import split_array_to_dict_of_arrays
 
 from gemseo_umdo.visualizations.uncertain_coupling_graph import UncertainCouplingGraph
+
+if TYPE_CHECKING:
+    from gemseo.problems.mdo.sobieski.disciplines import SobieskiDiscipline
 
 
 @pytest.fixture(scope="module")